move _to_object_array into validation
